// src/generated/rustplus.ts
export interface AppRequest {
  seq: number;
  playerId: string;
  playerToken: number;
  entityId?: number | undefined;
  getInfo?: AppEmpty | undefined;
  getTime?: AppEmpty | undefined;
  getMap?: AppEmpty | undefined;
  getTeamInfo?: AppEmpty | undefined;
  getTeamChat?: AppEmpty | undefined;
  sendTeamMessage?: AppSendMessage | undefined;
  getEntityInfo?: AppEmpty | undefined;
  setEntityValue?: AppSetEntityValue | undefined;
  checkSubscription?: AppEmpty | undefined;
  setSubscription?: AppFlag | undefined;
  getMapMarkers?: AppEmpty | undefined;
  promoteToLeader?: AppPromoteToLeader | undefined;
  getClanInfo?: AppEmpty | undefined;
  setClanMotd?: AppSendMessage | undefined;
  getClanChat?: AppEmpty | undefined;
  sendClanMessage?: AppSendMessage | undefined;
  getNexusAuth?: AppGetNexusAuth | undefined;
  cameraSubscribe?: AppCameraSubscribe | undefined;
  cameraUnsubscribe?: AppEmpty | undefined;
  cameraInput?: AppCameraInput | undefined;
}

export interface AppMessage {
  response?: AppResponse | undefined;
  broadcast?: AppBroadcast | undefined;
}

export interface AppResponse {
  seq: number;
  success?: AppSuccess | undefined;
  error?: AppError | undefined;
  info?: AppInfo | undefined;
  time?: AppTime | undefined;
  map?: AppMap | undefined;
  teamInfo?: AppTeamInfo | undefined;
  teamChat?: AppTeamChat | undefined;
  entityInfo?: AppEntityInfo | undefined;
  flag?: AppFlag | undefined;
  mapMarkers?: AppMapMarkers | undefined;
  clanInfo?: AppClanInfo | undefined;
  clanChat?: AppClanChat | undefined;
  nexusAuth?: AppNexusAuth | undefined;
  cameraSubscribeInfo?: AppCameraInfo | undefined;
}

export interface AppBroadcast {
  teamChanged?: AppTeamChanged | undefined;
  teamMessage?: AppNewTeamMessage | undefined;
  entityChanged?: AppEntityChanged | undefined;
  clanChanged?: AppClanChanged | undefined;
  clanMessage?: AppNewClanMessage | undefined;
  cameraRays?: AppCameraRays | undefined;
}

export interface AppEmpty {
}
// ...
export interface AppTeamInfo_Member {
  steamId: string;
  name: string;
  x: number;
  y: number;
  isOnline: boolean;
  spawnTime: number;
  isAlive: boolean;
  deathTime: number;
}
// ...
export const AppMessage: MessageFns<AppMessage> = {
  encode(message: AppMessage, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.response !== undefined) {
      AppResponse.encode(message.response, writer.uint32(10).fork()).join();
    }
    if (message.broadcast !== undefined) {
      AppBroadcast.encode(message.broadcast, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AppMessage {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAppMessage();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.response = AppResponse.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.broadcast = AppBroadcast.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<AppMessage>, I>>(base?: I): AppMessage {
    return AppMessage.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AppMessage>, I>>(object: I): AppMessage {
    const message = createBaseAppMessage();
    message.response = (object.response !== undefined && object.response !== null)
      ? AppResponse.fromPartial(object.response)
      : undefined;
    message.broadcast = (object.broadcast !== undefined && object.broadcast !== null)
      ? AppBroadcast.fromPartial(object.broadcast)
      : undefined;
    return message;
  },
};
// ...
export const AppBroadcast: MessageFns<AppBroadcast> = {
  encode(message: AppBroadcast, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.teamChanged !== undefined) {
      AppTeamChanged.encode(message.teamChanged, writer.uint32(34).fork()).join();
    }
    if (message.teamMessage !== undefined) {
      AppNewTeamMessage.encode(message.teamMessage, writer.uint32(42).fork()).join();
    }
    if (message.entityChanged !== undefined) {
      AppEntityChanged.encode(message.entityChanged, writer.uint32(50).fork()).join();
    }
    if (message.clanChanged !== undefined) {
      AppClanChanged.encode(message.clanChanged, writer.uint32(58).fork()).join();
    }
    if (message.clanMessage !== undefined) {
      AppNewClanMessage.encode(message.clanMessage, writer.uint32(66).fork()).join();
    }
    if (message.cameraRays !== undefined) {
      AppCameraRays.encode(message.cameraRays, writer.uint32(82).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AppBroadcast {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAppBroadcast();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.teamChanged = AppTeamChanged.decode(reader, reader.uint32());
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.teamMessage = AppNewTeamMessage.decode(reader, reader.uint32());
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.entityChanged = AppEntityChanged.decode(reader, reader.uint32());
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.clanChanged = AppClanChanged.decode(reader, reader.uint32());
          continue;
        }
        case 8: {
          if (tag !== 66) {
            break;
          }

          message.clanMessage = AppNewClanMessage.decode(reader, reader.uint32());
          continue;
        }
        case 10: {
          if (tag !== 82) {
            break;
          }

          message.cameraRays = AppCameraRays.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<AppBroadcast>, I>>(base?: I): AppBroadcast {
    return AppBroadcast.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AppBroadcast>, I>>(object: I): AppBroadcast {
    const message = createBaseAppBroadcast();
    message.teamChanged = (object.teamChanged !== undefined && object.teamChanged !== null)
      ? AppTeamChanged.fromPartial(object.teamChanged)
      : undefined;
    message.teamMessage = (object.teamMessage !== undefined && object.teamMessage !== null)
      ? AppNewTeamMessage.fromPartial(object.teamMessage)
      : undefined;
    message.entityChanged = (object.entityChanged !== undefined && object.entityChanged !== null)
      ? AppEntityChanged.fromPartial(object.entityChanged)
      : undefined;
    message.clanChanged = (object.clanChanged !== undefined && object.clanChanged !== null)
      ? AppClanChanged.fromPartial(object.clanChanged)
      : undefined;
    message.clanMessage = (object.clanMessage !== undefined && object.clanMessage !== null)
      ? AppNewClanMessage.fromPartial(object.clanMessage)
      : undefined;
    message.cameraRays = (object.cameraRays !== undefined && object.cameraRays !== null)
      ? AppCameraRays.fromPartial(object.cameraRays)
      : undefined;
    return message;
  },
};
// ...
export const AppTeamInfo_Member: MessageFns<AppTeamInfo_Member> = {
  encode(message: AppTeamInfo_Member, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.steamId !== "0") {
      writer.uint32(8).uint64(message.steamId);
    }
    if (message.name !== "") {
      writer.uint32(18).string(message.name);
    }
    if (message.x !== 0) {
      writer.uint32(29).float(message.x);
    }
    if (message.y !== 0) {
      writer.uint32(37).float(message.y);
    }
    if (message.isOnline !== false) {
      writer.uint32(40).bool(message.isOnline);
    }
    if (message.spawnTime !== 0) {
      writer.uint32(48).uint32(message.spawnTime);
    }
    if (message.isAlive !== false) {
      writer.uint32(56).bool(message.isAlive);
    }
    if (message.deathTime !== 0) {
      writer.uint32(64).uint32(message.deathTime);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AppTeamInfo_Member {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAppTeamInfo_Member();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.steamId = reader.uint64().toString();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.name = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 29) {
            break;
          }

          message.x = reader.float();
          continue;
        }
        case 4: {
          if (tag !== 37) {
            break;
          }

          message.y = reader.float();
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.isOnline = reader.bool();
          continue;
        }
        case 6: {
          if (tag !== 48) {
            break;
          }

          message.spawnTime = reader.uint32();
          continue;
        }
        case 7: {
          if (tag !== 56) {
            break;
          }

          message.isAlive = reader.bool();
          continue;
        }
        case 8: {
          if (tag !== 64) {
            break;
          }

          message.deathTime = reader.uint32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<AppTeamInfo_Member>, I>>(base?: I): AppTeamInfo_Member {
    return AppTeamInfo_Member.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AppTeamInfo_Member>, I>>(object: I): AppTeamInfo_Member {
    const message = createBaseAppTeamInfo_Member();
    message.steamId = object.steamId ?? "0";
    message.name = object.name ?? "";
    message.x = object.x ?? 0;
    message.y = object.y ?? 0;
    message.isOnline = object.isOnline ?? false;
    message.spawnTime = object.spawnTime ?? 0;
    message.isAlive = object.isAlive ?? false;
    message.deathTime = object.deathTime ?? 0;
    return message;
  },
};
// ...
export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}